+++ src/Libraries/RevitNodes/Elements/InternalUtilities/ElementUtils.cs
@@ -3,6 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Autodesk.DesignScript.Runtime;
+using Autodesk.Revit.DB;
+using RevitServices.Persistence;
+
 
 namespace Revit.Elements.InternalUtilities
 {
@@ -47,6 +50,66 @@ namespace Revit.Elements.InternalUtilities
             {
                 name += "(1)";
             }
+        }
+
+
+        /// <summary>
+        /// Get a revit parameters value
+        /// </summary>
+        /// <param name="param">Revit parameter</param>
+        /// <returns></returns>
+        public static object GetParameterValue(Autodesk.Revit.DB.Parameter param)
+        {
+            object result;
+
+            switch (param.StorageType)
+            {
+                case StorageType.ElementId:
+                    int valueId = param.AsElementId().IntegerValue;
+                    if (valueId > 0)
+                    {
+                        // When the element is obtained here, to convert it to our element wrapper, it
+                        // need to be figured out whether this element is created by us. Here the existing
+                        // element wrappers will be checked. If there is one, its property to specify
+                        // whether it is created by us will be followed. If there is none, it means the
+                        // element is not created by us.
+                        var elem = ElementIDLifecycleManager<int>.GetInstance().GetFirstWrapper(valueId) as Element;
+                        result = ElementSelector.ByElementId(valueId, elem == null ? true : elem.IsRevitOwned);
+                    }
+                    else
+                    {
+                        int paramId = param.Id.IntegerValue;
+                        if (paramId == (int)BuiltInParameter.ELEM_CATEGORY_PARAM || paramId == (int)BuiltInParameter.ELEM_CATEGORY_PARAM_MT)
+                        {
+                            var categories = DocumentManager.Instance.CurrentDBDocument.Settings.Categories;
+                            result = new Category(categories.get_Item((BuiltInCategory)valueId));
+                        }
+                        else
+                        {
+                            // For other cases, return a localized string
+                            result = param.AsValueString();
+                        }
+                    }
+                    break;
+                case StorageType.String:
+                    result = param.AsString();
+                    break;
+                case StorageType.Integer:
+                    result = param.AsInteger();
+                    break;
+                case StorageType.Double:
+                    var paramType = param.Definition.ParameterType;
+                    if (Element.IsConvertableParameterType(paramType))
+                        result = param.AsDouble() * Revit.GeometryConversion.UnitConverter.HostToDynamoFactor(
+                            Element.ParameterTypeToUnitType(paramType));
+                    else
+                        result = param.AsDouble();
+                    break;
+                default:
+                    throw new Exception(string.Format(Properties.Resources.ParameterWithoutStorageType, param));
+            }
+
+            return result;
         }
     }
 }
